requirements.txt
python-dotenv==0.21.*
pyyaml==6.0
pandas==1.5.*
psycopg2-binary==2.9.*
tenacity==8.1.*
pillow==9.3.*
pydantic==1.10.*
gspread==5.7.*
dash==2.7.*
dash-bootstrap-components==1.3.*
dash-html-components==2.0.*
dash_auth
Flask
sshtunnel==0.4.*
openpyxl
xlrd
xmltodict
kaleido
zeep
tqdm
typer
